+++ src/qasm/parser/parser.h
@@ -2,8 +2,8 @@
  * The parser of QuickAssembly. The inline assembly of Quickfall.
  */
 
-#include "../parser/structs/functions.h"
-#include "../ir/structs.h"
+#include "../../parser/structs/functions.h"
+#include "../../ir/structs.h"
 
 #ifndef QASM_PARSER_H
 #define QASM_PARSER_H
