+++ src/qasm/qasm.h
@@ -1,0 +1,10 @@
+/**
+ * The main file of QuickAssembly (QASm).
+ */
+
+#ifndef QASM_H
+#define QASM_H
+
+
+
+#endif
